Fix: Chip component replacment plus files restructuring
