+++ src/app/pages/PChainPages/index.tsx
@@ -1,4 +1,3 @@
 import PChainPage from './PChainPage';
-import PShowAllTransactions from './ShowAllTransactions';
 
-export { PChainPage, PShowAllTransactions };
+export { PChainPage };
