+++ src/app/pages/XChainPages/index.tsx
@@ -1,5 +1,5 @@
 import XChainPage from './XChainPage';
-import XAddressDetail from './XAddressDetail';
+import XAddressDetail from './Address/XAddressDetail';
 // import { XTransactionDetails } from './XTransactionDetails';
 import XPTransactions from './Transactions';
